moves ample filters out of public API
Ample filters were in a package that is in the public API. Since Ample
itself is not in the public API, these filters should also not be in the
public API.  This commit moves the filters to another packages that is
not in the public API
